Comment out autodie
Autodie would be good because the module does lots of system calls whose error codes are not checked. I wanted it to catch the bad system calls but actually IPC::System::Simple is not universally installed. My mistake for not provisioning it. Problem found by Yikun.

replace all system(...) with system(...) and die $? , and the module will work as I wanted it to work.